Host details from daahboard to hosts -jaminorja
